+++ LEDControl/StormSpirit.h
@@ -5,14 +5,14 @@
 #include "LEDMode.h"
 
 class StormSpirit : public LEDMode {
-    int delayCounter;
-    int count;
+    uint8_t delayCounter;
+    uint8_t count;
 
     public:
         StormSpirit() : delayCounter(0), count(0) {}
         void doLoop(){
             if (delayCounter == 0){
-                for (int i = 0; i < NUM_LEDS; i++){
+                for (uint8_t i = 0; i < NUM_LEDS; i++){
                     if      ((i + count) % 4 == 0)  leds[i] = ORANGE;
                     else if ((i + count) % 4 == 2)  leds[i] = RED;
                     else                            leds[i] = BLACK;
